+++ src/pygitgo/commands/git_core.py
@@ -1,6 +1,7 @@
 from pygitgo.auth.ssh_utils import convert_https_to_ssh, get_ssh_key_path, is_ssh_url, check_connection
 from pygitgo.utils.colors import info, success, warning, error
 from pygitgo.exceptions import GitGoError, GitCommandError
+from pygitgo.auth.account import sanitize_signing_config
 from pygitgo.commands.git_remote import handle_rebase
 from pygitgo.utils.config import get_default_branch
 from pygitgo.utils.executor import run_command
@@ -25,6 +26,8 @@ def git_commit(commit_message, loading_msg="Commiting changes...", skip_staging=
             return False
     except GitCommandError:
         return False
+    
+    sanitize_signing_config()
 
     if not skip_staging:
         run_command(["git", "add", "."], loading_msg="Staging files...")